+++ TidalPy/tides/multilayer/numerical_int/initial_conditions/initial_solution_dynamic_incomp.py
@@ -268,6 +268,7 @@ def solid_guess_takeuchi(
 
     # TODO
     raise Exception('Not Implemented for the Incompressible Assumption')
+    lame = 0.  # Just something temp so the linter does not break.
 
     # Convert compressibility parameters
     #lame = bulk_modulus - (2. / 3.) * shear_modulus
@@ -428,6 +429,7 @@ def liquid_guess_takeuchi(
 
     # TODO
     raise Exception('Not Implemented for the Incompressible Assumption')
+    lame = 0.  # Just something temp so the linter does not break.
 
     # Convert compressibility parameters
     # lame = bulk_modulus